Change content length's type to long

+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/contractimpl/HttpOutboundRespListener.java
@@ -61,7 +61,7 @@ public class HttpOutboundRespListener implements HttpConnectorListener {
     private ChunkConfig chunkConfig;
     private KeepAliveConfig keepAliveConfig;
     private boolean headerWritten = false;
-    private int contentLength = 0;
+    private long contentLength = 0;
     private String serverName;
     private List<HttpContent> contentList = new ArrayList<>();
