+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/listener/MaxEntityBodyValidator.java
@@ -41,12 +41,12 @@ public class MaxEntityBodyValidator extends ChannelInboundHandlerAdapter {
     private static final Logger log = LoggerFactory.getLogger(MaxEntityBodyValidator.class);
 
     private String serverName;
-    private int maxEntityBodySize;
-    private int currentSize;
+    private long maxEntityBodySize;
+    private long currentSize;
     private HttpRequest inboundRequest;
     private LinkedList<HttpContent> fullContent;
 
-    MaxEntityBodyValidator(String serverName, int maxEntityBodySize) {
+    MaxEntityBodyValidator(String serverName, long maxEntityBodySize) {
         this.serverName = serverName;
         this.maxEntityBodySize = maxEntityBodySize;
         this.fullContent = new LinkedList<>();
@@ -91,7 +91,7 @@ public class MaxEntityBodyValidator extends ChannelInboundHandlerAdapter {
         log.warn("Inbound request URI length exceeds the max uri length allowed for a request");
     }
 
-    private boolean isContentLengthInvalid(HttpMessage start, int maxContentLength) {
+    private boolean isContentLengthInvalid(HttpMessage start, long maxContentLength) {
         try {
             return HttpUtil.getContentLength(start, -1L) > (long) maxContentLength;
         } catch (NumberFormatException var4) {
